feat: Clean up UI, functionality, and convert to JS
