+++ esm-samples/jsapi-oauth/oauth-callback.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <script>
+      function loadHandler() {
+        if (opener) {
+          // opener.console.log("oauth callback href:", location.href);
+          if (location.hash) {
+            try {
+              var esriId = opener.require("esri/kernel").id;
+            } catch (e) {}
+            if (esriId) {
+              esriId.setOAuthResponseHash(location.hash);
+            } else {
+              opener.dispatchEvent(new CustomEvent("arcgis:auth:hash", { detail: location.hash }));
+            }
+          } else if (location.search) {
+            opener.dispatchEvent(new CustomEvent("arcgis:auth:location:search", { detail: location.search }));
+          }
+        }
+        close();
+      }
+    </script>
+  </head>
+  <body onload="loadHandler();"></body>
+</html>
